* Leaf fixes
svn: 1376

+++ src/mod/server.mod/servmsg.c
@@ -987,7 +987,7 @@ static void timeout_server(int idx)
   disconnect_server(idx, DO_LOST);
 }
 
-static void server_activity(int, char *, size_t);
+static void server_activity(int, char *, int);
 
 static struct dcc_table SERVER_SOCKET =
 {
@@ -1003,7 +1003,7 @@ static struct dcc_table SERVER_SOCKET =
   NULL
 };
 
-static void server_activity(int idx, char *msg, size_t len)
+static void server_activity(int idx, char *msg, int len)
 {
   char *from = NULL, *code = NULL;
